api/pyproject.toml
[project]
name = "voice-to-xero"
version = "0.1.0"
description = "Voice-to-Xero authentication and invoice creation tool for tradesmen"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "authlib>=1.2.1",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "itsdangerous>=2.1.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "email-validator>=2.3.0",
    "requests>=2.32.5",
    "xero-python>=9.0.0",
    "slowapi>=0.1.9",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "S", "C4", "PIE", "T20"]
ignore = ["E501", "S101"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.11",
]
